requirements.txt
# === Core Google AI & ADK ===
google-genai>=1.9.0
google-adk==0.1.0

# === Pydantic & Settings ===
pydantic>=2.5.0
pydantic-settings>=2.5.2
email-validator>=2.0.0

# === FastAPI + Server ===
fastapi==0.115.6
uvicorn[standard]>=0.34.0
python-multipart==0.0.6

# === Database & ORM ===
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# === Caching / Queue ===
redis==5.0.1
celery==5.3.4

# === Embeddings + Vector DB ===
chromadb==0.5.3
sentence-transformers==2.2.2

# === Async Networking ===
aiohttp==3.10.2
httpx>=0.27.0

# === Auth & Security ===
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === Observability ===
opentelemetry-api>=1.31.0
opentelemetry-sdk>=1.31.0
opentelemetry-instrumentation-fastapi>=0.42b0
prometheus-client==0.19.0

# === Utils ===
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3

# === Data Processing ===
numpy==1.24.3
pandas==2.0.3

# === Testing ===
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# === Dev Tools ===
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
